Moved database setup from DiscordMockingTestCase to subclass.
